feat(desktop): sync daemon sidecar, commands, and settings UI
- livtet-sync-daemon sidecar staged by build.rs and launched via tauri-plugin-shell
- sync_* commands proxying the daemon's stdio JSON-RPC; sync:// events
- settings route + sync panel (pairing, devices, conflicts, recent requests)
- remove livtet-plugin-host
- ADR 0012; glossary entries

+++ crates/livtet-desktop/src/commands/search.rs
@@ -51,7 +51,7 @@ pub async fn search_editions(
     let limit_usize = limit.unwrap_or(20).max(1) as usize;
 
     let opts = SearchOptions {
-        offset: offset_usize,
+        offset: offset_usize as i64,
         ..Default::default()
     };
